ph3

+++ src/components/TableData.jsx
@@ -18,6 +18,7 @@ import useTruncate from "../hooks/useTruncate";
 import { Link } from "react-router-dom";
 import { supabase } from "../utils/SupaClient";
 import Swal from "sweetalert2";
+import { useAuth } from "../auth/AuthProvider";
 
 // Daftar kolom untuk tabel
 const columns = [
@@ -47,27 +48,33 @@ const columns = [
   },
 ];
 
-export default function TablePaginate({ allBarang }) {
+export default function TablePaginate({ allBarang, search }) {
   const [page, setPage] = useState(1);
   const rowsPerPage = 5;
   const { formatRupiah } = useFormatRupiah();
   const [filterType, setFilterType] = useState("All");
 
+  const filteredItem = useMemo(() => {
+    return allBarang.filter((items) =>
+      items.name.toLowerCase().includes(search.toLowerCase())
+    );
+  }, [allBarang, search]);
+
   const { truncate } = useTruncate();
 
-  const filteredItems = useMemo(() => {
-    if (filterType === "All") {
-      return allBarang;
-    }
-    return allBarang.filter((item) => item.type === filterType);
-  }, [filterType, allBarang]);
+  // const filteredItems = useMemo(() => {
+  //   if (filterType === "All") {
+  //     return allBarang;
+  //   }
+  //   return allBarang.filter((item) => item.type === filterType);
+  // }, [filterType, allBarang]);
 
-  const pages = Math.ceil(filteredItems.length / rowsPerPage);
-  const items = useMemo(() => {
+  const pages = Math.ceil(filteredItem.length / rowsPerPage);
+  const items = React.useMemo(() => {
     const start = (page - 1) * rowsPerPage;
     const end = start + rowsPerPage;
-    return filteredItems.slice(start, end);
-  }, [page, filteredItems]);
+    return filteredItem.slice(start, end);
+  }, [page, filteredItem]);
 
   const deleteBarangById = async (id) => {
     try {
@@ -81,11 +88,42 @@ export default function TablePaginate({ allBarang }) {
         confirmButtonText: "Yes, delete it!",
       }).then(async (result) => {
         if (result.isConfirmed) {
-          const { data } = await supabase
+          const { data: getImageById } = await supabase
             .from("product")
-            .delete()
             .eq("id", id)
-            .select();
+            .select("image")
+            .single();
+
+          const removeUrlImage = String(getImageById.image).replace(
+            "https://gypbsaisblmgicbjarts.supabase.co/storage/v1/object/public/imageBuckets/items/",
+            ""
+          );
+
+          if (getImageById) {
+            const { data: removeImage } = await supabase.storage
+              .from("items")
+              .remove([`image/${removeUrlImage}`]);
+
+            if (removeImage) {
+              const { data } = await supabase
+                .from("product")
+                .delete()
+                .eq("id", id)
+                .select();
+
+              if (data) {
+                Swal.fire({
+                  title: "Are you sure?",
+                  text: "You won't be able to revert this!",
+                  icon: "warning",
+                  showCancelButton: true,
+                  confirmButtonColor: "#3085d6",
+                  cancelButtonColor: "#d33",
+                  confirmButtonText: "Yes, delete it!",
+                });
+              }
+            }
+          }
 
           if (data) {
             window.location.reload();
@@ -97,6 +135,7 @@ export default function TablePaginate({ allBarang }) {
     }
   };
 
+  const { user, role } = useAuth();
   return (
     <div>
       <div className="flex justify-between mb-4">
@@ -143,28 +182,33 @@ export default function TablePaginate({ allBarang }) {
                 <TableCell key={columnKey}>
                   {columnKey === "action" ? (
                     <div className="relative flex items-center gap-3">
-                      <Link to={`/detail/${item.id}`}>
-                        <Tooltip content="Detail">
-                          <span className="text-lg text-default-400 cursor-pointer active:opacity-50">
-                            <EyeIcon />
-                          </span>
-                        </Tooltip>
-                      </Link>
-                      <Link to={`/edit/${item.id}`}>
-                        <Tooltip content="Edit">
-                          <span className="text-lg text-default-400 cursor-pointer active:opacity-50">
-                            <EditIcon />
-                          </span>
-                        </Tooltip>
-                      </Link>
-                      <Tooltip color="danger" content="Delete">
-                        <span
-                          className="text-lg text-danger cursor-pointer active:opacity-50"
-                          onClick={() => deleteBarangById(item.id)}
-                        >
-                          <DeleteIcon />
-                        </span>
-                      </Tooltip>
+                      {user && role === "admin" ? (
+                        <>
+                          <Link to={`/edit/${item.id}`}>
+                            <Tooltip content="Edit">
+                              <span className="text-lg text-default-400 cursor-pointer active:opacity-50">
+                                <EditIcon />
+                              </span>
+                            </Tooltip>
+                          </Link>
+                          <Tooltip color="danger" content="Delete">
+                            <span
+                              className="text-lg text-danger cursor-pointer active:opacity-50"
+                              onClick={() => deleteBarangById(item.id)}
+                            >
+                              <DeleteIcon />
+                            </span>
+                          </Tooltip>
+                        </>
+                      ) : (
+                        <Link to={`/detail/${item.id}`}>
+                          <Tooltip content="Detail">
+                            <span className="text-lg text-default-400 cursor-pointer active:opacity-50">
+                              <EyeIcon />
+                            </span>
+                          </Tooltip>
+                        </Link>
+                      )}
                     </div>
                   ) : columnKey === "price" ? (
                     formatRupiah(getKeyValue(item, columnKey))
